edit get user

+++ TeamRedBackEnd/Controllers/UserController.cs
@@ -1,12 +1,7 @@
-﻿using Konscious.Security.Cryptography;
+﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Npgsql;
 using System;
-using System.Linq;
-using System.Net;
-using System.Net.Http;
-using System.Security.Cryptography;
-using System.Text;
+using System.Threading.Tasks;
 using TeamRedBackEnd.Database.Models;
 using TeamRedBackEnd.Database.Repositories;
 using TeamRedBackEnd.ViewModels;
@@ -15,17 +10,22 @@ namespace TeamRedBackEnd.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
+   // [Authorize]
+
     public class UserController : ControllerBase
     {
         private IUserRepository _userRepo;
         Services.PasswordService _passwordService;
         Services.IMailService _mailService;
 
+
+
         public UserController(IUserRepository userRepo, Services.PasswordService passwordService, Services.IMailService mailService)
         {
             _userRepo = userRepo;
             _passwordService = passwordService;
             _mailService = mailService;
+
         }
 
 
@@ -37,49 +37,42 @@ namespace TeamRedBackEnd.Controllers
 
 
         [HttpGet]
-        [Route("get/{id:int}")]
+        [Route("{id:int}")]
+
         public IActionResult GetUserById(int id)
         {
-            var user = _userRepo.GetUser(id);
-            if (user == null)
-            {
-                var response = new HttpResponseMessage(HttpStatusCode.NotFound)
-                {
-                    ReasonPhrase = "User with ID: " + id + " doesn't exist"
-                };
 
-                return NotFound(response.ReasonPhrase);
-            }
+            var user = _userRepo.GetUser(id);
+
+            if (user == null) return NotFound("User with ID: " + id + " doesn't exist");
+         
             return Ok(user);
+
         }
 
 
+
         [HttpGet]
-        [Route("get/{name}")]
+        [Route("name/{name}")]
+        [AllowAnonymous]
         public IActionResult GetUserByNameOrEmail(string name)
         {
             if (String.IsNullOrEmpty(name))
             {
                 return NotFound("Input can't be null");
             }
+            var user = _userRepo.GetUser(name);
 
-            else
+            return user switch
             {
-                try
-                {
-                    var user = _userRepo.GetUser(name);
-                    return Ok(user);
-                }
-                catch
-                {
-                    return NotFound("No user found with this name: " + name);
-                }
-            }
+                null => NotFound("No user found with this name: " + name),
+                _ => Ok(user),
+            };
         }
 
 
         [HttpPost]
-        [Route("create")]
+        [AllowAnonymous]
         public IActionResult AddNewUser([FromBody] Usermodel usermodel)
         {
 
@@ -94,7 +87,7 @@ namespace TeamRedBackEnd.Controllers
                 return BadRequest(ModelState);
             }
 
-            else if (_userRepo.GetUser(usermodel.Email) != null)
+            if (_userRepo.GetUser(usermodel.Email) != null)
             {
                 ModelState.AddModelError("email", "Email address already in use");
                 return BadRequest(ModelState);
@@ -109,33 +102,31 @@ namespace TeamRedBackEnd.Controllers
             MailRequest mail = _mailService.MakeVerificationMail(usermodel);
             _mailService.SendMailAsync(mail);
 
+
             return Ok("User " + usermodel.Name + " has been created\n");
         }
 
 
-        [HttpPatch]
+        [HttpGet]
         [Route("{VerificationLinkCode}")]
-        public IActionResult VerifyUser([FromBody]string VerificationCode)
+        public IActionResult VerifyUser(string VerificationLinkCode)
         {
-            User user = _userRepo.GetUserByVerificationCode(VerificationCode);
+            User user = _userRepo.GetUserByVerificationCode(VerificationLinkCode);
             if (user == null) return BadRequest("Invalid verification link");
             user.Verified = true;
             _userRepo.EditUser(user);
+
             return Ok(user);
         }
 
 
         [HttpPatch]
-        [Route("edit/{id:int}")]
+        [Route("{id:int}")]
         public ActionResult<Usermodel> EditUserProfile(int id, [FromBody] Usermodel usermodel)
         {
             var existingUserProfile = _userRepo.GetUser(id);
-
             var checkIfNameExists = _userRepo.GetUser(usermodel.Name);
-
             var checkIfEmailExists = _userRepo.GetUser(usermodel.Email);
-
-
             _passwordService.CreateSalt(usermodel);
 
             if (!ModelState.IsValid) return BadRequest(ModelState);
@@ -155,53 +146,50 @@ namespace TeamRedBackEnd.Controllers
 
             usermodel.Id = existingUserProfile.Id;
 
-            
             _passwordService.HashPassword(usermodel);
             _userRepo.EditUser(usermodel);
 
-
-            return Ok("User has been edited");
+            return Ok(usermodel);
         }
 
 
         [HttpDelete]
-        [Route("delete/{id:int}")]
+        [Route("{id:int}")]
         public IActionResult DeleteUser(int id)
         {
-            try
+            var userToDelete = _userRepo.GetUser(id);
+            ObjectResult objectresult = NotFound("User with ID: " + id + " doesn't exist");
+
+            if (userToDelete != null)
             {
-                var userToDelete = _userRepo.GetUser(id);
+                objectresult = Ok("User has been deleted");
                 _userRepo.RemoveUser(userToDelete);
-                return Ok("User has been deleted");
+
             }
-            catch
-            {
-                return NotFound("User with ID: " + id + " doesn't exist");
-            }
+
+            return objectresult;
+
         }
 
 
         [HttpDelete]
-        [Route("delete/{name}")]
+        [Route("{name}")]
         public IActionResult DeleteUser(string name)
         {
             if (String.IsNullOrWhiteSpace(name))
             {
                 return NotFound("Input can't be null");
             }
+            var userToDelete = _userRepo.GetUser(name);
 
-            else
+            switch (userToDelete)
             {
-                try
-                {
-                    var userToDelete = _userRepo.GetUser(name);
+                case null:
+                    return NotFound("No user found with this name: " + name);
+
+                default:
                     _userRepo.RemoveUser(userToDelete);
                     return Ok("User has been deleted");
-                }
-                catch
-                {
-                    return NotFound("No user found with this name: " + name);
-                }
             }
         }
